Add GetKudos HTTP method

+++ kudo.go
@@ -1,8 +1,8 @@
 package main
 
 type Kudo struct {
-	ID         int64  `json:"id"`
-	SenderID   int64  `json:"senderId"`
-	ReceiverID int64  `json:"receiverId"`
+	ID         int32  `json:"id"`
+	SenderID   int32  `json:"senderId"`
+	ReceiverID int32  `json:"receiverId"`
 	Message    string `json:"message"`
 }
